+++ validation/certificates/k3s/cert_rotation_test.go
@@ -1,12 +1,11 @@
-//go:build validation || recurring
+//go:build (validation || infra.rke2k3s || recurring || cluster.any || stress) && !infra.any && !infra.aks && !infra.eks && !infra.gke && !infra.rke1 && !sanity && !extended
 
-package dualstack
+package k3s
 
 import (
 	"os"
 	"testing"
 
-	rkev1 "github.com/rancher/rancher/pkg/apis/rke.cattle.io/v1"
 	"github.com/rancher/shepherd/clients/rancher"
 	v1 "github.com/rancher/shepherd/clients/rancher/v1"
 	extClusters "github.com/rancher/shepherd/extensions/clusters"
@@ -14,15 +13,14 @@ import (
 	"github.com/rancher/shepherd/pkg/config"
 	"github.com/rancher/shepherd/pkg/config/operations"
 	"github.com/rancher/shepherd/pkg/session"
+	"github.com/rancher/tests/actions/certificates"
 	"github.com/rancher/tests/actions/clusters"
 	"github.com/rancher/tests/actions/config/defaults"
 	"github.com/rancher/tests/actions/logging"
 	"github.com/rancher/tests/actions/provisioning"
-	"github.com/rancher/tests/actions/provisioninginput"
 	"github.com/rancher/tests/actions/qase"
 	"github.com/rancher/tests/actions/workloads/deployment"
 	"github.com/rancher/tests/actions/workloads/pods"
-	"github.com/rancher/tests/validation/certificates"
 	resources "github.com/rancher/tests/validation/provisioning/resources/provisioncluster"
 	standard "github.com/rancher/tests/validation/provisioning/resources/standarduser"
 	"github.com/sirupsen/logrus"
@@ -30,20 +28,19 @@ import (
 	"github.com/stretchr/testify/suite"
 )
 
-type CertRotationDualstackTestSuite struct {
+type CertRotationTestSuite struct {
 	suite.Suite
 	session      *session.Session
 	client       *rancher.Client
 	cattleConfig map[string]any
-	rke2Cluster  *v1.SteveAPIObject
-	k3sCluster   *v1.SteveAPIObject
+	cluster      *v1.SteveAPIObject
 }
 
-func (c *CertRotationDualstackTestSuite) TearDownSuite() {
+func (c *CertRotationTestSuite) TearDownSuite() {
 	c.session.Cleanup()
 }
 
-func (c *CertRotationDualstackTestSuite) SetupSuite() {
+func (c *CertRotationTestSuite) SetupSuite() {
 	testSession := session.NewSession()
 	c.session = testSession
 
@@ -72,54 +69,55 @@ func (c *CertRotationDualstackTestSuite) SetupSuite() {
 	provider := provisioning.CreateProvider(clusterConfig.Provider)
 	machineConfigSpec := provider.LoadMachineConfigFunc(c.cattleConfig)
 
-	logrus.Info("Provisioning RKE2 cluster")
-	c.rke2Cluster, err = resources.ProvisionRKE2K3SCluster(c.T(), standardUserClient, extClusters.RKE2ClusterType.String(), provider, *clusterConfig, machineConfigSpec, nil, true, false)
-	require.NoError(c.T(), err)
+	rancherConfig := new(rancher.Config)
+	operations.LoadObjectFromMap(defaults.RancherConfigKey, c.cattleConfig, rancherConfig)
 
-	if clusterConfig.Advanced == nil {
-		clusterConfig.Advanced = &provisioninginput.Advanced{}
+	if rancherConfig.ClusterName == "" {
+		logrus.Info("Provisioning K3S cluster")
+		c.cluster, err = resources.ProvisionRKE2K3SCluster(c.T(), standardUserClient, extClusters.K3SClusterType.String(), provider, *clusterConfig, machineConfigSpec, nil, true, false)
+		require.NoError(c.T(), err)
+	} else {
+		logrus.Infof("Using existing cluster %s", rancherConfig.ClusterName)
+		c.cluster, err = c.client.Steve.SteveType(stevetypes.Provisioning).ByID("fleet-default/" + c.client.RancherConfig.ClusterName)
+		require.NoError(c.T(), err)
 	}
 
-	if clusterConfig.Advanced.MachineGlobalConfig == nil {
-		clusterConfig.Advanced.MachineGlobalConfig = &rkev1.GenericMap{
-			Data: map[string]any{},
-		}
-	}
-
-	clusterConfig.Advanced.MachineGlobalConfig.Data["flannel-ipv6-masq"] = true
-
-	logrus.Info("Provisioning K3s cluster")
-	c.k3sCluster, err = resources.ProvisionRKE2K3SCluster(c.T(), standardUserClient, extClusters.K3SClusterType.String(), provider, *clusterConfig, machineConfigSpec, nil, true, false)
-	require.NoError(c.T(), err)
 }
 
-func (c *CertRotationDualstackTestSuite) TestCertRotationDualstack() {
+func (c *CertRotationTestSuite) TestCertRotation() {
 	tests := []struct {
-		name      string
-		clusterID string
+		name    string
+		cluster *v1.SteveAPIObject
 	}{
-		{"RKE2_Dualstack_Certificate_Rotation", c.rke2Cluster.ID},
-		{"K3S_Dualstack_Certificate_Rotation", c.k3sCluster.ID},
+		{"K3S_Certificate_Rotation", c.cluster},
 	}
 
 	for _, tt := range tests {
-		cluster, err := c.client.Steve.SteveType(stevetypes.Provisioning).ByID(tt.clusterID)
-		require.NoError(c.T(), err)
-
+		var err error
 		c.Run(tt.name, func() {
-			logrus.Infof("Rotating certificates on cluster (%s)", cluster.Name)
-			require.NoError(c.T(), certificates.RotateCerts(c.client, cluster.Name))
-
-			logrus.Infof("Verifying the cluster is ready (%s)", cluster.Name)
-			provisioning.VerifyClusterReady(c.T(), c.client, cluster)
-
-			logrus.Infof("Verifying cluster deployments (%s)", cluster.Name)
-			err = deployment.VerifyClusterDeployments(c.client, cluster)
+			oldCertificates, err := certificates.GetClusterCertificates(c.client, tt.cluster.Name)
 			require.NoError(c.T(), err)
 
-			logrus.Infof("Verifying cluster pods (%s)", cluster.Name)
-			err = pods.VerifyClusterPods(c.client, cluster)
+			logrus.Infof("Rotating certificates on cluster (%s)", tt.cluster.Name)
+			require.NoError(c.T(), certificates.RotateCerts(c.client, tt.cluster.Name))
+
+			logrus.Infof("Verifying the cluster is ready (%s)", tt.cluster.Name)
+			provisioning.VerifyClusterReady(c.T(), c.client, tt.cluster)
+
+			logrus.Infof("Verifying cluster deployments (%s)", tt.cluster.Name)
+			err = deployment.VerifyClusterDeployments(c.client, tt.cluster)
 			require.NoError(c.T(), err)
+
+			logrus.Infof("Verifying cluster pods (%s)", tt.cluster.Name)
+			err = pods.VerifyClusterPods(c.client, tt.cluster)
+			require.NoError(c.T(), err)
+
+			newCertificates, err := certificates.GetClusterCertificates(c.client, tt.cluster.Name)
+			require.NoError(c.T(), err)
+
+			logrus.Infof("Verifying certificates were rotated (%s)", tt.cluster.Name)
+			isRotated := certificates.VerifyCertificateRotation(oldCertificates, newCertificates)
+			require.True(c.T(), isRotated)
 		})
 
 		params := provisioning.GetProvisioningSchemaParams(c.client, c.cattleConfig)
@@ -130,6 +128,6 @@ func (c *CertRotationDualstackTestSuite) TestCertRotationDualstack() {
 	}
 }
 
-func TestCertRotationDualstackTestSuite(t *testing.T) {
-	suite.Run(t, new(CertRotationDualstackTestSuite))
+func TestCertRotationTestSuite(t *testing.T) {
+	suite.Run(t, new(CertRotationTestSuite))
 }
