+++ validation/certificates/k3s/dualstack/defaults/defaults.yaml
