+++ validation/certificates/k3s/ipv6/defaults/defaults.yaml
