+++ validation/certificates/k3s/ipv6/schemas/hostbusters_schemas.yaml
@@ -1,0 +1,23 @@
+- suite: Go Automation/Certificates/k3s/ipv6
+  projects: [RRT, RM]
+  cases:
+  - description: Rotates the clusters certificates
+    title: K3S_IPv6_Certificate_Rotation
+    priority: 4
+    type: 8
+    is_flaky: 0
+    automation: 2
+    steps:
+    - action: Rotate the clusters certificates
+      expectedresult: ""
+      data: ""
+      position: 1
+      attachments: []
+    - action: Verify cluster state
+      expectedresult: ""
+      data: ""
+      position: 2
+      attachments: []
+    custom_field:
+      "14": Validation
+      "18": Hostbusters
