+++ validation/certificates/rke2/dualstack/defaults/defaults.yaml
@@ -1,0 +1,13 @@
+#Required for all dualstack/ipv6 tests
+clusterConfig:
+  networking:
+    clusterCIDR: "<required>"
+    serviceCIDR: "<required>"
+    stackPreference: "<required>"
+
+awsMachineConfigs:
+  awsMachineConfig:
+  - enablePrimaryIPv6: true
+    httpProtocolIpv6: "enabled"
+    ipv6AddressOnly: true
+    ipv6AddressCount: "1"
