--- validation/certificates/rke2k3s/cert_rotation_existing_cluster_test.go
@@ -1,95 +1,0 @@
-//go:build (validation || infra.rke2k3s || cluster.any || stress) && !infra.any && !infra.aks && !infra.eks && !infra.gke && !infra.rke1 && !sanity && !extended
-
-package rke2k3s
-
-import (
-	"os"
-	"testing"
-
-	"github.com/rancher/shepherd/clients/rancher"
-	v1 "github.com/rancher/shepherd/clients/rancher/v1"
-	"github.com/rancher/shepherd/extensions/defaults/stevetypes"
-	"github.com/rancher/shepherd/pkg/config"
-	"github.com/rancher/shepherd/pkg/config/operations"
-	"github.com/rancher/shepherd/pkg/session"
-	"github.com/rancher/tests/actions/config/defaults"
-	"github.com/rancher/tests/actions/logging"
-	"github.com/rancher/tests/actions/provisioning"
-	"github.com/rancher/tests/actions/workloads/deployment"
-	"github.com/rancher/tests/actions/workloads/pods"
-	"github.com/rancher/tests/validation/certificates"
-	"github.com/sirupsen/logrus"
-	"github.com/stretchr/testify/require"
-	"github.com/stretchr/testify/suite"
-)
-
-type CertRotationExistingClusterTestSuite struct {
-	suite.Suite
-	session       *session.Session
-	client        *rancher.Client
-	cattleConfig  map[string]any
-	clusterObject *v1.SteveAPIObject
-}
-
-func (c *CertRotationExistingClusterTestSuite) TearDownSuite() {
-	c.session.Cleanup()
-}
-
-func (c *CertRotationExistingClusterTestSuite) SetupSuite() {
-	testSession := session.NewSession()
-	c.session = testSession
-
-	client, err := rancher.NewClient("", testSession)
-	require.NoError(c.T(), err)
-
-	c.client = client
-
-	c.cattleConfig = config.LoadConfigFromFile(os.Getenv(config.ConfigEnvironmentKey))
-
-	c.cattleConfig, err = defaults.LoadPackageDefaults(c.cattleConfig, "")
-	require.NoError(c.T(), err)
-
-	loggingConfig := new(logging.Logging)
-	operations.LoadObjectFromMap(logging.LoggingKey, c.cattleConfig, loggingConfig)
-
-	err = logging.SetLogger(loggingConfig)
-	require.NoError(c.T(), err)
-
-	c.clusterObject, err = client.Steve.SteveType(stevetypes.Provisioning).ByID("fleet-default/" + c.client.RancherConfig.ClusterName)
-	require.NoError(c.T(), err)
-
-}
-
-func (c *CertRotationExistingClusterTestSuite) TestCertRotationExistingCluster() {
-	tests := []struct {
-		name      string
-		clusterID string
-	}{
-		{"ExistingCluster_Certificate_Rotation", c.clusterObject.ID},
-	}
-
-	for _, tt := range tests {
-		cluster, err := c.client.Steve.SteveType(stevetypes.Provisioning).ByID(tt.clusterID)
-		require.NoError(c.T(), err)
-
-		c.Run(tt.name, func() {
-			logrus.Infof("Rotating certificates on cluster (%s)", cluster.Name)
-			require.NoError(c.T(), certificates.RotateCerts(c.client, cluster.Name))
-
-			logrus.Infof("Verifying the cluster is ready (%s)", cluster.Name)
-			provisioning.VerifyClusterReady(c.T(), c.client, cluster)
-
-			logrus.Infof("Verifying cluster deployments (%s)", cluster.Name)
-			err = deployment.VerifyClusterDeployments(c.client, cluster)
-			require.NoError(c.T(), err)
-
-			logrus.Infof("Verifying cluster pods (%s)", cluster.Name)
-			err = pods.VerifyClusterPods(c.client, cluster)
-			require.NoError(c.T(), err)
-		})
-	}
-}
-
-func TestCertRotationExistingClusterTestSuite(t *testing.T) {
-	suite.Run(t, new(CertRotationExistingClusterTestSuite))
-}
